removes UMD wrapper, moves lib/argon2.js to index.js
