--- lib/argon2.js
@@ -1,159 +1,0 @@
-!function(globalRoot, factory) {
-  if ('function' === typeof define && define.amd)
-    define('mymodule', [], function() { return (globalRoot.mymodule = factory()) })
-  else if ('object' === typeof module && module.exports)
-    module.exports = factory()
-  else
-    globalRoot.mymodule = factory()
-}(this, function() {
-  'use strict'
-
-  var isBrowser = typeof window !== 'undefined';
-  var g = isBrowser ? window : global;
-  var root = isBrowser ? '' : '../';
-
-  /**
-   * Argon2 hash
-   * @param {string} params.pass - password string
-   * @param {string} params.salt - salt string
-   * @param {float}  [params.time=1] - the number of iterations
-   * @param {float}  [params.mem=1024] - used memory, in KiB
-   * @param {float}  [params.hashLen=24] - desired hash length
-   * @param {float}  [params.parallelism=1] - desired parallelism (will be computed in parallel only for PNaCl)
-   * @param {number} [params.type=argon2.ArgonType.Argon2d] - hash type: argon2.ArgonType.Argon2d, .Argon2i, .Argon2id or .Argon2u
-   * @param {number} [params.distPath=.] - asm.js script location, without trailing slash
-   *
-   * @return Promise
-   *
-   * @example
-   *  argon2.hash({ pass: 'password', salt: 'somesalt' })
-   *      .then(h => console.log(h.hash, h.hashHex, h.encoded))
-   *      .catch(e => console.error(e.message, e.code))
-   */
-  function argon2(args) {
-      if (!WebAssembly) {
-          return new Promise((resolve, reject) => {
-              reject({message: 'WebAssembly not supported here.', code:-4040});
-          });
-      }
-
-      if (g.Module && g.Module._argon2_hash) {
-          return new Promise((resolve, reject) => {
-              try {
-                  resolve(calcHash(args));
-              } catch(e) {
-                  reject(e);
-              }
-          });
-      }
-
-      const loadModule = require('../generated/argon2.js')
-
-      // TODO: this is unused
-      // should it be set in g.Module?
-
-      // const KB = 1024 * 1024;
-      // const MB = 1024 * KB;
-      // const GB = 1024 * MB;
-      // const WASM_PAGE_SIZE = 64 * 1024;
-
-      // const totalMemory = (2*GB - 64*KB) / 1024 / WASM_PAGE_SIZE;
-      // const mem = args.mem || +(1024);
-      // const initialMemory = Math.min(Math.max(Math.ceil(mem * 1024 / WASM_PAGE_SIZE), 256) + 256, totalMemory);
-      // const wasmMemory = new WebAssembly.Memory({
-      //     initial: initialMemory,
-      //     maximum: totalMemory
-      // });
-
-      g.Module = {
-          print: console.log,
-          printErr: console.error,
-          setStatus: console.log,
-          wasmBinary: null,
-          // wasmBinaryFile: root + 'dist/argon2.wasm',
-          // locateFile: function(file) { return (args.distPath || '') + '/' + file; }
-      };
-
-      return new Promise((resolve, reject) => {
-          g.Module.onRuntimeInitialized = function() {
-              try {
-                  resolve(calcHash(args));
-              } catch(e) {
-                  reject(e);
-              }
-          };
-
-          g.Module['wasmBinary'] = require('../generated/argon2.wasm.js');
-          loadModule(g.Module);
-      }).catch(err => {
-        console.log(err)
-      });
-  }
-
-  function calcHash(arg) {
-      if (!Module._argon2_hash) {
-          return console.error('Error: _argon2_hash not available');
-      }
-      var t_cost = arg && arg.time || 10;
-      var m_cost = arg && arg.mem || 1024;
-      var parallelism = arg && arg.parallelism || 1;
-      var pwd = allocateArray(arg && arg.pass || 'password');
-      var pwdlen = arg && arg.pass ? arg.pass.length : 8;
-      var salt = allocateArray(arg && arg.salt || 'somesalt');
-      var saltlen = arg && arg.salt ? arg.salt.length : 8;
-      var hash = Module.allocate(new Array(arg && arg.hashLen || 32), 'i8', Module.ALLOC_NORMAL);
-      var hashlen = arg && arg.hashLen || 32;
-      var encoded = Module.allocate(new Array(512), 'i8', Module.ALLOC_NORMAL);
-      var encodedlen = 512;
-      var argon2_type = arg && arg.type || 0;
-      var version = 0x13;
-      var err;
-      var out = false;
-      try {
-          var res = Module._argon2_hash(t_cost, m_cost, parallelism, pwd, pwdlen, salt, saltlen,
-              hash, hashlen, encoded, encodedlen,
-              argon2_type, version);
-      } catch (e) {
-          err = e;
-      }
-      var result;
-      if (res === 0 && !err) {
-          var hashStr = '';
-          var hashArr = new Uint8Array(hashlen);
-          for (var i = 0; i < hashlen; i++) {
-              var byte = Module.HEAP8[hash + i];
-              hashArr[i] = byte;
-              hashStr += ('0' + (0xFF & byte).toString(16)).slice(-2);
-          }
-          var encodedStr = Module.Pointer_stringify(encoded);
-          result = { hash: hashArr, hashHex: hashStr, encoded: encodedStr };
-      } else {
-          try {
-              if (!err) {
-                  err = Module.Pointer_stringify(Module._argon2_error_message(res))
-              }
-          } catch (e) {
-          }
-          result = { message: err, code: res };
-      }
-      try {
-          Module._free(pwd);
-          Module._free(salt);
-          Module._free(hash);
-          Module._free(encoded);
-      } catch (e) { }
-      if (err) {
-          throw result;
-      } else {
-          return result;
-      }
-  }
-
-  function allocateArray(strOrArr) {
-      var arr = strOrArr instanceof Uint8Array || strOrArr instanceof Array ? strOrArr
-          : Module.intArrayFromString(strOrArr);
-      return Module.allocate(arr, 'i8', Module.ALLOC_NORMAL);
-  }
-
-  return argon2;
-})
